first extract of common code from Pdo & Dbal
